HashMap Implemented
HashSet was the wrong tool; we needed a mapping function that acts like
a HashTable which is HashMap. I implemented this into MaxMove and
MinMove, but have yet to figure how to coordinate the Iterative part of
the IDS. So far, I can only get the best scores, but have no way of
knowing which move is associated with that score.